In GEM geometry, use GemDetRing instead of MuDetRing
GemDetRing will fall back on a full phi search in the case that the
2018 slice test geometry has been setup. The assumption of equally
phi-spaced detectors used in the tracking search breaks down when in
the partial GEM slice test.

Also, one change to order the Gem eta partitions correctly by r (since
the partition number is reversed with respect to r, high partitions
are at low r).